requirements-dev.txt
Django==5.0.7
django-environ>=0.11.2
python-slugify==8.0.4
python3-openid==3.2.0
psycopg2-binary==2.9.6
requests==2.32.3
requests-oauthlib==2.0.0
djangorestframework
python-docx
PyPDF2

